Set up <C-Space> and <leader><CR>
<C-Space> is not as easy to map to as I thought. To get it one should
use <C-@>. In general one can press <C-V> in the insert mode and then
the desired button combo to see how to map to it (also lets you easy put
symbols for stuff like enter or escape).

Reverted back to leader <CR>, because both <C-CR> and <S-CR> are impossible to
map to.